overlays: perf_metrics_overlay to seperate header
this is done to prevent severe conflicts with upcoming changes
